DSS-2747 : avoid float conversion from COSNumber class (preserve original numeric type)
(cherry picked from commit aabf1bf10630e23d8cdd97e598e5bbab64fc25f8)

+++ dss-pades-openpdf/src/main/java/eu/europa/esig/dss/pdf/openpdf/ITextPdfDict.java
@@ -138,7 +138,7 @@ class ITextPdfDict implements eu.europa.esig.dss.pdf.PdfDict {
 	public Number getNumberValue(String name) {
 		PdfNumber pdfNumber = wrapped.getAsNumber(new PdfName(name));
 		if (pdfNumber != null) {
-			return pdfNumber.floatValue();
+			return pdfNumber.doubleValue();
 		}
 		return null;
 	}
